# walk: Use unbounded channels