@tavianator 2023-11-01
> Wow 🤩

:)

> I would like to understand this first before merging. The situation is the following: we have a MPSC scenario where the consumer is sometimes too slow to handle the incoming results (even though its only job is to print the paths to a console). In the past, we used an unbounded channel which would lead to high memory usage because the channel was buffering all those results.

Exactly.  It was actually fairly easy to cause this by stalling the receiver thread with something like `fd | less`.

> Then we switched to a bounded channel, but that came at a high initialization cost because the bounded channels pre-allocate a fixed size buffer for the messages. The messages are `WorkerResult`s (`size_of::<WorkerResult>() == 312`). This fixed-size buffer presumably has a size of `channel_size × message_size`, i.e. `0x4000 × 312 byte ≈ 4.9 MiB` (?).

Yeah that times the number of threads: https://github.com/sharkdp/fd/blob/15329f9cfad43a1d9bef96914e9e30af5f68eae8/src/walk.rs#L60

> And that was slowing us down by ~70 ms on your machine? Is it because each thread allocates those 5 MiB?

I'm not sure exactly why `bounded()` channels are that slow to initialize, but the allocation happens up-front, not in each thread, which is part of the problem.

> In this changeset, you switch back to an unbounded channel, but `WorkerResult`s are re-cycled. I understand that this is getting rid of the pre-allocation hit. But what about long searches (when initialization time can be neglected)? You seem to indicate that we are still faster in those situations?

Yeah in my benchmarks it's universally faster.  But I don't know the separate impacts of

- Switching to an unbounded channel
- Shrinking `WorkerResult` to `WorkerMsg`
- Re-using `Box<WorkerResult>` allocations

> Couldn't this optimization be applied to bounded crossbeam channels in general (for large message sizes)? Is the allocation/deallocation really that expensive that it is worth all of this overhead (creating a sender for each worker result, making an additional copy for each worker result when recycling, …)?

Actually `Sender::clone()` is fairly cheap, it just [bumps an atomic refcount](https://github.com/crossbeam-rs/crossbeam/blob/17fecd9a442ab1582f75c81ce945bc4a66dc5c1e/crossbeam-channel/src/counter.rs#L47-L61).  I think it would even be possible to do

```rust
pub struct WorkerMsg<'a> {
    inner: Option<ResultBox>,
    tx: &'a Sender<ResultBox>,
}
```

but I'd have to plumb the lifetime through more stuff.  `WorkerState` would have to own the `Sender`s I think.

> Don't get me wrong. I love that this works. I'm just a bit puzzled that this isn't a strategy that could be used to speed up bounded channels in general (or maybe it is?).

Maybe it's possible?  Keep in mind this is somewhat different semantics because each thread is strictly limited to its pool of 0x4000 `WorkerResult`.  In the previous implementation, the capacity was shared between threads.

> Would another allocator help? I'm not really knowledgeable here but it seems to me like this whole recycling-memory part should/could be the job of a (special-purpose) allocator?

It's possible that https://github.com/microsoft/snmalloc would have some of the same benefits.  But here's the thing: this approach needs the receiver thread to tell the sender thread that it has handled a `WorkerResult` somehow.  My first thought was to use a semaphore to limit the number of `WorkerResult`s allocated.  But a single semaphore doesn't scale very well, so then I figured I could have a separate semaphore for each sender.

But actually, an SPSC channel scales better than a semaphore!  That's because the head/tail index can be on separate cache lines.  For a semaphore, both threads are always contending on the same counter.  So for that reason I just went with the channel implementation (and because we already had it available; I'd have to go find/write a semaphore otherwise).