@tavianator 2023-11-01
> I think it would even be possible to do
> 
> ```rust
> pub struct WorkerMsg<'a> {
>     inner: Option<ResultBox>,
>     tx: &'a Sender<ResultBox>,
> }
> ```
> 
> but I'd have to plumb the lifetime through more stuff. `WorkerState` would have to own the `Sender`s I think.

Update: I just tried a hacky version of this and it wasn't appreciably faster