@tmccombs 2023-11-02
> Re-using Box<WorkerResult> allocations

I suspect this is a pretty big component of it.  Unless the worker generates results faster than the receiver can process them, and we get to the maximum amount of allocation. 

It probably also helps that allocations for the channels now happens on the individual threads instead of in the spawning thread. 

Looking at the code for crossbeam, it isn't just that we have to allocate the memory for the channel up front, we also have to[ _initialize_ it](https://github.com/crossbeam-rs/crossbeam/blob/9dd4c9594dd7c84abf04e89192f1bdfe790100f5/crossbeam-channel/src/flavors/array.rs#L108), and it can't be done with `memset` or equivalent. 

> I'm just a bit puzzled that this isn't a strategy that could be used to speed up bounded channels in general (or maybe it is?)

I think it could be used in general. Basically implement a bounded channel as two unbounded channels and an atomic counter for the number of items allocated. And have the receiver pass the slot back after reading the value out of it. I'm not sure if that would universally improve performance, but it does have a couple of advantages: memory is allocated lazily, and it would be possible to dynamically change the size of the bound.