@tavianator 2023-11-02
> Looking at the code for crossbeam, it isn't just that we have to allocate the memory for the channel up front, we also have to[ _initialize_ it](https://github.com/crossbeam-rs/crossbeam/blob/9dd4c9594dd7c84abf04e89192f1bdfe790100f5/crossbeam-channel/src/flavors/array.rs#L108), and it can't be done with `memset` or equivalent.

Well it *could* be done with `memset()` if they changed their representation slightly, e.g.

```rust
/// A slot in a channel.
struct Slot<T> {
    /// The current stamp.
    stamp: AtomicUsize,
    /// The message in this slot.
    msg: UnsafeCell<MaybeUninit<T>>,
}

impl<T> Slot<T> {
    fn new() -> Self {
        stamp: AtomicUsize::new(),
        msg: UnsafeCell::new(MaybeUninit::zeroed()),
    }
}
...
let slot = unsafe { self.buffer.get_unchecked(index) };
let stamp = slot.stamp.load(Ordering::Acquire) + index;
```

And if they did that they could allocate the whole buffer with `mmap()` and get lazily-initialized zero pages.  I use a similar trick in bfs to initialize a whole linked list from zeroed memory: https://github.com/tavianator/bfs/blob/b2ab7a151fca517f4879e76e626ec85ad3de97c7/src/alloc.c#L63-L88