@sharkdp 2023-11-02
I did some benchmarks comparing current master (8bbbd7679b7eaf8782559c0320fc6cc86314319f) with this branch (d5889712454979f1f9ac4174b9f725135401b560).

I can confirm the huge increase in startup speed (`hyperfine -w5 -N -L version master,1414 "./fd-{version} -u . /tmp/empty" --export-markdown -`)

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `./fd-master -u . /tmp/empty` | 25.4 ± 10.0 | 18.1 | 54.7 | 7.24 ± 3.34 |
| `./fd-1414 -u . /tmp/empty` | 3.5 ± 0.9 | 2.6 | 9.5 | 1.00 |

Unfortunately, there seems to be a large regression for longer searches with many results (`hyperfine -w5 -N -L version master,1414 "./fd-{version} -u . /folder/with/3M/files" --export-markdown -`):

| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|:---|---:|---:|---:|---:|
| `./fd-master -u . /home/ped1st/workspace` | 841.0 ± 125.2 | 737.2 | 1069.5 | 1.00 |
| `./fd-1414 -u . /home/ped1st/workspace` | 1386.5 ± 18.0 | 1358.8 | 1411.5 | 1.65 ± 0.25 |